# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PriceHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listMyProducts
      tags: [products]
      summary: List the current seller's products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductCreate" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /marketplace/search:
    get:
      operationId: searchMarketplace
      tags: [marketplace]
      summary: Search products and compare prices across Russian marketplaces
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MarketplaceListing" }
...
  /marketplace/listings/{id}:
    get:
      operationId: getListing
      tags: [marketplace]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketplaceListing" }
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      responses:
        "200":
          description: Chats
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatThread" }
    post:
      operationId: createChat
      tags: [chats]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatCreate" }
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatThread" }
  /chats/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chats]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
    post:
      operationId: sendChatMessage
      tags: [chats]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageBody" }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Role:
      type: string
      enum: [buyer, seller, developer, support]
...
    Product:
      type: object
      required: [id, sellerId, name, price, cost, marketplace, currency, stock, createdAt, updatedAt]
      properties:
        id: { type: integer }
        sellerId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        cost: { type: number }
        currency: { type: string }
        marketplace: { type: string }
        category: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        externalUrl: { type: ["string", "null"] }
        stock: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductCreate:
      type: object
      required: [name, price, cost, marketplace]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        cost: { type: number }
        currency: { type: string }
        marketplace: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        externalUrl: { type: string }
        stock: { type: integer }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        cost: { type: number }
        currency: { type: string }
        marketplace: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        externalUrl: { type: string }
        stock: { type: integer }
    MarketplaceOffer:
      type: object
      required: [marketplace, price, url, inStock]
      properties:
        marketplace: { type: string }
        price: { type: number }
        url: { type: string }
        inStock: { type: boolean }
        deliveryDays: { type: integer }
        rating: { type: number }
    MarketplaceListing:
      type: object
      required: [id, name, category, imageUrl, lowestPrice, highestPrice, currency, offers]
      properties:
        id: { type: integer }
        name: { type: string }
        brand: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        category: { type: string }
        imageUrl: { type: string }
        currency: { type: string }
        lowestPrice: { type: number }
        highestPrice: { type: number }
        savingsPercent: { type: number }
        offers:
          type: array
          items: { $ref: "#/components/schemas/MarketplaceOffer" }
...
    ContactUser:
      type: object
      required: [id, hubId, displayName, role]
      properties:
        id: { type: integer }
        hubId: { type: string }
        displayName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        avatarUrl: { type: ["string", "null"] }
    ChatThread:
      type: object
      required: [id, otherUser, lastMessageAt, unread]
      properties:
        id: { type: integer }
        otherUser: { $ref: "#/components/schemas/ContactUser" }
        lastMessage: { type: ["string", "null"] }
        lastMessageAt: { type: string, format: date-time }
        unread: { type: integer }
    ChatMessage:
      type: object
      required: [id, chatId, senderId, content, sentAt]
      properties:
        id: { type: integer }
        chatId: { type: integer }
        senderId: { type: integer }
        senderName: { type: string }
        content: { type: string }
        sentAt: { type: string, format: date-time }
    ChatCreate:
      type: object
      required: [otherUserId]
      properties:
        otherUserId: { type: integer }
    SendMessageBody:
      type: object
      required: [content]
      properties:
        content: { type: string }
...